+++ package.json
@@ -15,7 +15,7 @@
     "axios": "^0.24.0",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
-    "react-hook-form": "^7.22.5",
+    "react-hook-form": "^7.23.0",
     "react-input-mask": "^2.0.4",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^4.0.0",
